refactor: split UIRenderSystem into focused systems following ECS
Refactored ui_render.py (468 lines) into two focused systems to follow
Single Responsibility Principle and keep systems under 300 lines.

Changes:
- Split UIRenderSystem into two systems:
  - UIRenderSystem: basic HUD (score, speed bar, music indicator) - 259 lines
  - OverlayRenderSystem: overlays (pause, settings menus) - 246 lines
- Updated GameplayScene to use both systems
- Each system now has clear, single responsibility
- Follows ECS coding standards for system size and focus

Both systems are now under the 300-line guideline and maintain proper
separation of concerns.
