+++ src/ecs/systems/ui_render.py
@@ -17,10 +17,10 @@
 #   You should have received a copy of the GNU General Public License
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""UIRenderSystem - handles rendering of UI overlays (score, speed bar, music indicator).
+"""UIRenderSystem - handles rendering of basic HUD elements.
 
 This system follows ECS Single Responsibility Principle by handling ONLY
-UI element rendering on top of the game world.
+basic HUD element rendering (score, speed bar, music indicator).
 """
 
 import pygame
@@ -33,15 +33,15 @@ from src.game import constants
 
 
 class UIRenderSystem(BaseSystem):
-    """System responsible for rendering UI overlays.
+    """System responsible for rendering basic HUD elements.
 
     Responsibilities:
     - Render score counter
     - Render speed bar
     - Render music indicator
-    - Render pause overlay (when paused)
 
     NOT responsible for:
+    - Overlays (use OverlayRenderSystem)
     - Game world rendering (use BoardRenderSystem)
     - Entity rendering (use EntityRenderSystem, SnakeRenderSystem)
     """
@@ -64,12 +64,12 @@ class UIRenderSystem(BaseSystem):
             surface_width: Width of the surface
             surface_height: Height of the surface
         """
-        # Query score component from world
+        # query score component from world
         score_entities = world.registry.query_by_component("score")
         if not score_entities:
             return
 
-        # Get first score entity
+        # get first score entity
         score_entity = list(score_entities.values())[0]
         if not hasattr(score_entity, "score"):
             return
@@ -77,7 +77,7 @@ class UIRenderSystem(BaseSystem):
         current_score = score_entity.score.current
 
         try:
-            # Large font size
+            # large font size
             font_size = int(surface_width / 8)
             font_path = "assets/font/GetVoIP-Grotesque.ttf"
 
@@ -86,27 +86,27 @@ class UIRenderSystem(BaseSystem):
             except Exception:
                 score_font = pygame.font.Font(None, font_size)
 
-            # Get color from constants
+            # get color from constants
             score_color = Color.from_hex(constants.MESSAGE_COLOR).to_tuple()
 
-            # Render score text
+            # render score text
             score_text = score_font.render(str(current_score), True, score_color)
 
-            # Make it translucent (~25% opaque)
+            # make it translucent (~25% opaque)
             score_text.set_alpha(64)
 
-            # Horizontal center; vertically near the top with margin
+            # horizontal center; vertically near the top with margin
             top_margin = getattr(
                 world.board, "cell_size", max(10, surface_height // 20)
             )
             score_rect = score_text.get_rect()
             score_rect.midtop = (surface_width // 2, top_margin)
 
-            # Blit to main surface
+            # blit to main surface
             self._renderer.blit(score_text, score_rect)
 
         except Exception:
-            # Silently fail if font loading or rendering fails
+            # silently fail if font loading or rendering fails
             pass
 
     def draw_speed_bar(
@@ -122,7 +122,7 @@ class UIRenderSystem(BaseSystem):
         if not self._settings:
             return
 
-        # Get speed settings
+        # get speed settings
         initial_speed = self._settings.get("initial_speed")
         max_speed = self._settings.get("max_speed")
         if initial_speed is None or max_speed is None:
@@ -131,7 +131,7 @@ class UIRenderSystem(BaseSystem):
         min_speed = float(initial_speed)
         max_speed = float(max_speed)
 
-        # Get current speed from snake
+        # get current speed from snake
         snakes = world.registry.query_by_type(EntityType.SNAKE)
         current_speed = min_speed
         for _, snake in snakes.items():
@@ -139,20 +139,20 @@ class UIRenderSystem(BaseSystem):
                 current_speed = snake.velocity.speed
                 break
 
-        # Geometry
+        # geometry
         padding_x = int(surface_width * 0.02)
         padding_y = int(surface_height * 0.02)
         bar_width = int(surface_width * 0.25)
         bar_height = int(surface_height * 0.02)
         gap = 6
 
-        # Colors - bar changes from green (slow) to red (fast)
+        # colors - bar changes from green (slow) to red (fast)
         if max_speed > min_speed:
             ratio = (current_speed - min_speed) / (max_speed - min_speed)
         else:
             ratio = 0.0
         ratio = max(0.0, min(ratio, 1.0))
-        # Bar color changes from green (slow) to red (fast) - calculated dynamically
+        # bar color changes from green (slow) to red (fast) - calculated dynamically
         bar_color = (
             int(255 * ratio),
             int(255 * (1 - ratio)),
@@ -161,24 +161,24 @@ class UIRenderSystem(BaseSystem):
         border_color = Color.from_hex(constants.GRID_COLOR).to_tuple()
         text_color = Color.from_hex(constants.MESSAGE_COLOR).to_tuple()
 
-        # Bar position
+        # bar position
         bar_x = padding_x
         bar_y = padding_y
 
-        # Create temporary surface for the speed bar
+        # create temporary surface for the speed bar
         bar_surface = pygame.Surface((bar_width, bar_height))
         bar_surface.fill(border_color)
 
-        # Draw filled portion
+        # draw filled portion
         filled_width = int(bar_width * ratio)
         if filled_width > 0:
             filled_rect = pygame.Rect(0, 0, filled_width, bar_height)
             pygame.draw.rect(bar_surface, bar_color, filled_rect)
 
-        # Blit bar to screen
+        # blit bar to screen
         self._renderer.blit(bar_surface, (bar_x, bar_y))
 
-        # Draw text label below
+        # draw text label below
         label_text = f"Speed: {current_speed:.1f}"
         font_size = int(surface_width / 50)
         font_path = "assets/font/GetVoIP-Grotesque.ttf"
@@ -192,7 +192,7 @@ class UIRenderSystem(BaseSystem):
         label_rect = label_surf.get_rect()
         label_rect.midtop = (bar_x + bar_width // 2, bar_y + bar_height + gap)
 
-        # Blit label
+        # blit label
         self._renderer.blit(label_surf, label_rect)
 
     def draw_music_indicator(
@@ -206,13 +206,13 @@ class UIRenderSystem(BaseSystem):
             music_on: Whether background music is currently enabled
         """
         try:
-            # Define dimensions
+            # define dimensions
             padding_x = int(surface_width * 0.02)
             padding_y = int(surface_height * 0.02)
             icon_size = int(surface_width / 25)
             gap = 4
 
-            # Load speaker sprites
+            # load speaker sprites
             try:
                 if music_on:
                     sprite = pygame.image.load("assets/sprites/speaker-on.png")
@@ -221,7 +221,7 @@ class UIRenderSystem(BaseSystem):
             except Exception:
                 sprite = None
 
-            # Render hint text - white when on, dim grid color when off
+            # render hint text - white when on, dim grid color when off
             hint_color = (
                 Color.from_hex(constants.SCORE_COLOR).to_tuple()
                 if music_on
@@ -239,217 +239,36 @@ class UIRenderSystem(BaseSystem):
             hint_surf = hint_font.render(hint_text, True, hint_color)
             hint_rect = hint_surf.get_rect()
 
-            # Calculate total widget height
+            # calculate total widget height
             total_widget_height = icon_size + gap + hint_rect.height
 
-            # Calculate positions (bottom-right corner)
+            # calculate positions (bottom-right corner)
             icon_x = surface_width - padding_x - icon_size
             icon_y = surface_height - padding_y - total_widget_height
 
-            # Scale and draw sprite
+            # scale and draw sprite
             if sprite is not None:
                 scaled_sprite = pygame.transform.scale(sprite, (icon_size, icon_size))
                 self._renderer.blit(scaled_sprite, (icon_x, icon_y))
 
-            # Position and draw text hint below the icon
+            # position and draw text hint below the icon
             hint_rect.centerx = icon_x + icon_size // 2
             hint_rect.top = icon_y + icon_size + gap
             self._renderer.blit(hint_surf, hint_rect)
 
         except Exception:
-            # Silently fail if sprite loading or rendering fails
-            pass
-
-    def draw_pause_overlay(self, surface_width: int, surface_height: int) -> None:
-        """Draw pause overlay with semi-transparent background and text.
-
-        Args:
-            surface_width: Width of the surface
-            surface_height: Height of the surface
-        """
-        try:
-            # Create semi-transparent overlay
-            overlay = pygame.Surface((surface_width, surface_height))
-            overlay.set_alpha(128)  # 50% transparent
-            overlay.fill(Color.from_hex(constants.ARENA_COLOR).to_tuple())
-            self._renderer.blit(overlay, (0, 0))
-
-            # Render "PAUSED" text
-            font_size = int(surface_width / 10)
-            font_path = "assets/font/GetVoIP-Grotesque.ttf"
-
-            try:
-                pause_font = pygame.font.Font(font_path, font_size)
-            except Exception:
-                pause_font = pygame.font.Font(None, font_size)
-
-            pause_text = pause_font.render(
-                "PAUSED", True, Color.from_hex(constants.SCORE_COLOR).to_tuple()
-            )
-            pause_rect = pause_text.get_rect()
-            pause_rect.center = (surface_width // 2, surface_height // 2)
-
-            # Blit to screen
-            self._renderer.blit(pause_text, pause_rect)
-
-            # Render hint text below
-            hint_font_size = int(surface_width / 30)
-            try:
-                hint_font = pygame.font.Font(font_path, hint_font_size)
-            except Exception:
-                hint_font = pygame.font.Font(None, hint_font_size)
-
-            hint_text = hint_font.render(
-                "Press P to resume or ESC/M for settings",
-                True,
-                Color.from_hex(constants.MESSAGE_COLOR).to_tuple(),
-            )
-            hint_rect = hint_text.get_rect()
-            hint_rect.midtop = (surface_width // 2, pause_rect.bottom + 20)
-
-            # Blit hint
-            self._renderer.blit(hint_text, hint_rect)
-
-        except Exception:
-            # Silently fail if rendering fails
-            pass
-
-    def draw_settings_overlay(
-        self,
-        surface_width: int,
-        surface_height: int,
-        selected_index: int,
-        config=None,
-    ) -> None:
-        """Draw settings overlay with semi-transparent background and settings menu.
-
-        Args:
-            surface_width: Width of the surface
-            surface_height: Height of the surface
-            selected_index: Currently selected setting index
-            config: Game config for calculating grid size
-        """
-        if not self._settings:
-            return
-
-        try:
-            # Create semi-transparent overlay
-            overlay = pygame.Surface((surface_width, surface_height))
-            overlay.set_alpha(200)  # More opaque than pause
-            overlay.fill(Color.from_hex(constants.ARENA_COLOR).to_tuple())
-            self._renderer.blit(overlay, (0, 0))
-
-            # Draw title
-            font_path = "assets/font/GetVoIP-Grotesque.ttf"
-            title_font_size = int(surface_width / 12)
-            try:
-                title_font = pygame.font.Font(font_path, title_font_size)
-            except Exception:
-                title_font = pygame.font.Font(None, title_font_size)
-
-            title_text = title_font.render(
-                "Settings", True, Color.from_hex(constants.MESSAGE_COLOR).to_tuple()
-            )
-            title_rect = title_text.get_rect(
-                center=(surface_width / 2, surface_height / 10)
-            )
-            self._renderer.blit(title_text, title_rect)
-
-            # Spacing and scroll parameters
-            row_h = int(surface_height * 0.06)
-            visible_rows = int(surface_height * 0.70 // row_h)
-            top_index = max(0, selected_index - visible_rows + 3)
-            padding_y = int(surface_height * 0.22)
-
-            # Draw visible rows - only show in-game adjustable settings
-            menu_fields = self._settings.get_in_game_menu_fields()
-            return_to_menu_index = len(menu_fields)
-
-            item_font_size = int(surface_width / 30)
-            try:
-                item_font = pygame.font.Font(font_path, item_font_size)
-            except Exception:
-                item_font = pygame.font.Font(None, item_font_size)
-
-            # Draw settings items
-            for draw_i, field_i in enumerate(range(top_index, len(menu_fields))):
-                if draw_i >= visible_rows:
-                    break
-                f = menu_fields[field_i]
-                val = self._settings.get(f["key"])
-
-                # Calculate current grid size for display
-                current_grid_size = 20  # default fallback
-                if config:
-                    desired_cells = max(10, int(self._settings.get("cells_per_side")))
-                    current_grid_size = config.get_optimal_grid_size(desired_cells)
-
-                formatted_val = self._settings.format_setting_value(
-                    f,
-                    val,
-                    surface_width,
-                    current_grid_size,
-                )
-
-                # Render text with highlighting for selected item
-                text_color = (
-                    Color.from_hex(constants.SCORE_COLOR).to_tuple()
-                    if field_i == selected_index
-                    else Color.from_hex(constants.MESSAGE_COLOR).to_tuple()
-                )
-                text = item_font.render(
-                    f"{f['label']}: {formatted_val}", True, text_color
-                )
-                rect = text.get_rect()
-                rect.left = int(surface_width * 0.10)
-                rect.top = padding_y + draw_i * row_h
-                self._renderer.blit(text, rect)
-
-            # Draw "Return to Menu" option
-            return_draw_i = len(menu_fields) - top_index
-            if return_draw_i >= 0 and return_draw_i < visible_rows:
-                text_color = (
-                    Color.from_hex(constants.SCORE_COLOR).to_tuple()
-                    if selected_index == return_to_menu_index
-                    else Color.from_hex(constants.MESSAGE_COLOR).to_tuple()
-                )
-                separator_top = padding_y + return_draw_i * row_h
-                # add some spacing before the option
-                return_text = item_font.render("Return to Main Menu", True, text_color)
-                rect = return_text.get_rect()
-                rect.left = int(surface_width * 0.10)
-                rect.top = separator_top + int(row_h * 0.5)  # add spacing
-                self._renderer.blit(return_text, rect)
-
-            # Hint footer
-            hint_text = "[A/D] change   [W/S] navigate   [Enter] select   [Esc] back   [C] random colors"
-            hint_font_size = int(surface_width / 50)
-            try:
-                hint_font = pygame.font.Font(font_path, hint_font_size)
-            except Exception:
-                hint_font = pygame.font.Font(None, hint_font_size)
-
-            hint_surf = hint_font.render(
-                hint_text, True, Color.from_hex(constants.GRID_COLOR).to_tuple()
-            )
-            hint_rect = hint_surf.get_rect(
-                center=(surface_width / 2, surface_height * 0.95)
-            )
-            self._renderer.blit(hint_surf, hint_rect)
-
-        except Exception:
-            # Silently fail if rendering fails
+            # silently fail if sprite loading or rendering fails
             pass
 
     def update(self, world: World) -> None:
         """Update method required by BaseSystem.
 
-        Renders UI overlays on top of the game world.
+        Renders basic HUD elements on top of the game world.
 
         Args:
             world: Game world to render
         """
-        # Get surface dimensions
+        # get surface dimensions
         surface = pygame.display.get_surface()
         if not surface:
             return
@@ -457,11 +276,11 @@ class UIRenderSystem(BaseSystem):
         surface_width = surface.get_width()
         surface_height = surface.get_height()
 
-        # Draw UI elements
+        # draw UI elements
         self.draw_score(world, surface_width, surface_height)
         self.draw_speed_bar(world, surface_width, surface_height)
 
-        # Draw music indicator
+        # draw music indicator
         if self._settings:
             music_on = self._settings.get("background_music")
             self.draw_music_indicator(surface_width, surface_height, music_on)
